Fixed error with timeout (to) as reported at http://seclists.org/nmap-dev/2010/q2/450 . It appears that nsock_gettimeofday() was the culprit as nsock doesn't record the time as frequently and as a result, the now time was a bit behind than the real one. This led to a negative value as a result of the timeval subtraction.

+++ Service.h
@@ -244,7 +244,7 @@ class Service
     /* number of connection retries after connection failure */
 		long connection_retries;
     /* maximum cracking time regardless of success so far */
-    long timeout;
+    long long timeout;
 
 		/* misc options */
 		bool ssl;   /* true -> SSL enabled over this service */
